rag-support-assistant/requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-google-genai==2.0.0
langgraph==0.2.40
chromadb==0.5.5
sentence-transformers==3.0.1
pypdf==5.0.0
python-dotenv==1.0.1